Add hash function and regression tests
This function is based off of [`hash_array`][1], which is a built-in
PostgreSQL function for producing a hash value from the elements of an
array of a single type. Currently this function operates on a single
attribute only, but by adding a loop similar to the one in `hash_array`
it could be augmented to support hash keys composed of values from many
columns (unlike `hash_array`, they needn't be of the same type).

The logic also resembles that of `worker_hash_partition_table` within
CitusDB, ensuring compatibility with the repartitioning functions that
live there.

**Caveat**: If PostgreSQL's included hash functions change, the output
of this function will also change. If this function is used to provide
shard placements for distributed data, those placements may be invalid
after a change in hash functions. So upgrade with care.

[1]: https://github.com/postgres/postgres/blob/REL9_3_4/src/backend/utils/adt/arrayfuncs.c#L3528-L3642