Update .gitignore for standalone module
PostgreSQL has a top-level `.gitignore` file to ensure build artifacts
are not committed. Since we've ripped out this subdirectory, we need an
updated `.gitignore` specific to this project.

The C ignores come from GitHub's [gitignore project][1]. Ignores that
are specific to this project go in the bottom section. `pg_regress`
unfortunately uses the `out` suffix for its output, so we add a rule to
"unignore" the `.out` files we care about.

[1]: https://github.com/github/gitignore