Establish shard connections in BeginForeignModify
Augments `PgFdwModifyState` to have an array of `TopsieShardConnInfo`
pointers, each of which contains information about the min/max values
in a shard as well as connections to nodes of that shard (if needed).

Currently, just establishes connections to every node during the begin
foreign modify phase.

Also contains code (as yet unused) to build shard-specific prepared
statements by appending a separator and shard number to the remote
relation name.