Add helpers to get partition Var and hash function
The function to return the partition column is pretty straightforward:
it returns the `Var` representing a partition column of a particular
foreign relation. This comes from the relation's `partition_column`
`OPTION`.

The `Var` references the `Oid` of a row in `pg_type`, which can then
be used to quickly find the hash function for a particular type. To
avoid repeated lookups, we store a `FunctionCallInfo` in the state
for the foreign modification. During the modification itself, a value
should be populated in this structure in order to allow the function
itself to be called.