Add connection hash functionality
Heavily influenced (or copied outright) from postgres_fdw's own file
named connection.c. I initially had more functionality related to
transactionality (which was the part most heavily modified as compared
to the connection.c in postgres_fdw), but decided it is not appropriate
in this file: consistency/atomicity will be handled in a separate file.