Attempt to clean up error reporting method
Decided the logic to call errmsg instead of errmsg_internal was not
necessary here, so assigned a default message to the primaryMessage
variable and just pass that to errmsg_internal. Also used some (within
CitusDB) unconventional spacing to help clean up the ternaries. Think
we want to keep them, since this is the main error reporting function
so having as much information as possible is desirable.