Add bad connection pruning callback
This handles cleanup of connections in bad states during transaction
event processing (pre-commit, commit, prepare, abort, etc.). I'm not
100% certain it's safe to stash away a pointer to a hash entry, but if
it is this should be an ideal solution, even if tens of thousands of
connections are in the hash: we'll only ever iterate over those that
were used in a particular transaction.