Mapping and pragmas for include-what-you-use
IWYU is useful for cleaning up includes. It has some false positives,
though they can be fixed with pragma comments. Additionally, it has
trouble finding the canonical location for certain library functions
and makes suggestions that are wrong unless you tell it where to find
the "right" version of a declaration (often important on platforms that
map certain functions to "secure" internal implementations).